Check note: expanded from macro
