+++ clang/test/Frontend/macro_defined_type.cpp
@@ -1,6 +1,4 @@
-// RUN: %clang_cc1 -I%S/Inputs -fsyntax-only -verify -triple x86_64-linux-gnu %s
-
-#include <macro_defined_type.h>
+// RUN: %clang_cc1 -fsyntax-only -verify -triple x86_64-linux-gnu %s
 
 #define NODEREF __attribute__((noderef))
 
@@ -16,24 +14,10 @@ void Func() {
   auto NODEREF auto_i2 = i; // expected-warning{{'noderef' can only be used on an array or pointer type}}
 }
 
-// The diagnostic message is hard-coded as 'noderef' so using a system macro doesn't change the behavior
-void Func_system_macro() {
-  int _SYS_NODEREF i; // expected-warning{{'noderef' can only be used on an array or pointer type}}
-  int _SYS_NODEREF *i_ptr;
-
-  auto _SYS_NODEREF *auto_i_ptr2 = i_ptr;
-  auto _SYS_NODEREF auto_i2 = i; // expected-warning{{'noderef' can only be used on an array or pointer type}}
-}
-
-
 // Added test for fix for P41835
 #define _LIBCPP_FLOAT_ABI __attribute__((pcs("aapcs")))
 struct A {
   _LIBCPP_FLOAT_ABI int operator()() throw(); // expected-warning{{'pcs' calling convention is not supported for this target}}
-};
-
-struct A_system_macro {
-  _SYS_LIBCPP_FLOAT_ABI int operator()() throw(); // expected-warning{{'_SYS_LIBCPP_FLOAT_ABI' calling convention is not supported for this target}}
 };
 
 // Added test for fix for PR43315
